@Polsaker 2021-06-12
Ohh those are some pretty good optimizations, good job!